+++ api.bak/v1/views/places.py
