--- api/__init__.py
@@ -1,2 +1,0 @@
-#!/usr/bin/python3
-"""Initializer Module"""
