+++ api/v1/views/__init__.py
@@ -1,15 +1,8 @@
 #!/usr/bin/python3
-""" Views initialization module """
+"""This module initializes the views"""
+
 from flask import Blueprint
 
-
 app_views = Blueprint('app_views', __name__, url_prefix='/api/v1')
-
-if app_views is not None:
+if app_views:
     from api.v1.views.index import *
-    from api.v1.views.states import *
-    from api.v1.views.cities import *
-    from api.v1.views.amenities import *
-    from api.v1.views.users import *
-    from api.v1.views.places import *
-    from api.v1.views.places_reviews import *
