+++ file.json
@@ -1,1 +1,1 @@
-{"State.acf25d1c-1e45-4c0a-9324-e02cc5e893ce": {"id": "acf25d1c-1e45-4c0a-9324-e02cc5e893ce", "created_at": "2024-02-26T11:31:00.287046", "updated_at": "2024-02-26T11:31:00.287094", "__class__": "State"}, "City.a11ddc39-24e0-4bf6-a8f1-fa682fa2230d": {"id": "a11ddc39-24e0-4bf6-a8f1-fa682fa2230d", "created_at": "2024-02-26T11:31:03.458508", "updated_at": "2024-02-26T11:31:03.458532", "__class__": "City"}}
+{"Amenity.289a5485-87f5-42e4-b95c-2118e9b3be3b": {"id": "289a5485-87f5-42e4-b95c-2118e9b3be3b", "created_at": "2024-02-28T12:01:18.087181", "updated_at": "2024-02-28T12:01:18.087181", "__class__": "Amenity"}, "BaseModel.76a1937e-1e3c-4c11-bfdb-bdcb1d6e23b1": {"id": "76a1937e-1e3c-4c11-bfdb-bdcb1d6e23b1", "created_at": "2024-02-28T12:01:18.087188", "updated_at": "2024-02-28T12:01:18.087188", "__class__": "BaseModel"}, "City.54354f45-fb95-48ea-b161-ed8150bf7b7d": {"id": "54354f45-fb95-48ea-b161-ed8150bf7b7d", "created_at": "2024-02-28T12:01:18.087194", "updated_at": "2024-02-28T12:01:18.087194", "__class__": "City"}, "Place.8487b2ad-0feb-4728-8bf4-fd028ef9054e": {"id": "8487b2ad-0feb-4728-8bf4-fd028ef9054e", "created_at": "2024-02-28T12:01:18.087200", "updated_at": "2024-02-28T12:01:18.087200", "__class__": "Place"}, "Review.44092a20-d570-43ad-8fc0-fc8470ca2a9c": {"id": "44092a20-d570-43ad-8fc0-fc8470ca2a9c", "created_at": "2024-02-28T12:01:18.087206", "updated_at": "2024-02-28T12:01:18.087206", "__class__": "Review"}, "State.3e88d451-0a3e-48cb-a57f-e7ecad483dc0": {"id": "3e88d451-0a3e-48cb-a57f-e7ecad483dc0", "created_at": "2024-02-28T12:01:18.087211", "updated_at": "2024-02-28T12:01:18.087211", "__class__": "State"}, "User.5ea1c0fe-9dc2-4821-b587-499082e107b1": {"id": "5ea1c0fe-9dc2-4821-b587-499082e107b1", "created_at": "2024-02-28T12:01:18.087217", "updated_at": "2024-02-28T12:01:18.087217", "__class__": "User"}}
